Changed ProductType id to ne type Long instead to int

+++ src/main/java/com/waldronprojects/bookstore/entity/ProductType.java
@@ -9,7 +9,7 @@ public class ProductType {
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name="name")
     private String name;
@@ -25,17 +25,17 @@ public class ProductType {
         this.description = description;
     }
 
-    public ProductType(int id, String name, String description) {
+    public ProductType(Long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
